Añadidos módulos
Añadidos los módulos para registro, login y admin. Se ha implementado las funcionalidades para el de registro, el resto no.
