+++ templates/index3.html
@@ -30,7 +30,7 @@
 
 
 <!-- Primer bloque -->
-<!-- CAMBIO DE COLOR  style="color:white" --> 
+<!-- CAMBIO DE COLOR  style="color:white" -->
 <div class="row vh-100"  style="color:white;height:100%;">
   <section class="container-fluid container-main" >
     <div class="row mb-5">
@@ -61,7 +61,7 @@
                 <div>
                   <div><h1 class = "titulo"> DISEÑO DIGITAL </h1> </div>
                 </div>
-                
+
           		</div>
 
             <div><h1 id="tituloPermanente" class="titulo" style="display:none">COMPANY DAY</h1></div>
